+++ src/selector/components/index.ts
@@ -1,2 +1,3 @@
 export * from './selector/selector.component';
 export * from './inputDateTime/inputDateTime.component';
+export * from './simpleInputDateTime/simpleInputDateTime.component';