+++ src/selector/index.ts
@@ -1,4 +1,5 @@
 export * from './modules/selector.module';
+export * from './modules/basicSelector.module';
 export * from './components';
 export * from './directives';
 export * from './misc/datetimeSelector.interface';