+++ src/selector/modules/basicSelector.module.ts
@@ -1,0 +1,17 @@
+import {NgModule} from '@angular/core';
+
+import {SimpleInputDateTimeSelectorComponent} from '../components';
+
+/**
+ * Angular module for date time basic input selector component
+ */
+@NgModule(
+{
+    declarations:
+    [
+        SimpleInputDateTimeSelectorComponent
+    ]
+})
+export class DateTimeBasicSelectorModule
+{
+}